% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature-responce-curve.R
\name{define_temperature_ranges}
\alias{define_temperature_ranges}
\title{Define Temperature Ranges}
\usage{
define_temperature_ranges(temperatures, n = 100)
}
\arguments{
\item{temperatures}{A numeric vector containing temperature values.}

\item{n}{An integer specifying the length of the output sequence. Default is 100.}
}
\value{
A numeric vector containing a sequence of temperature values ranging from the minimum temperature
to the maximum temperature plus 1, evenly spaced based on the specified length.
}
\description{
This function takes a vector of temperatures and generates a sequence of temperature values that span the range of input temperatures.
}
\examples{
temperatures <- c(10, 15, 20, 25, 30)
define_temperature_ranges(temperatures)
}
