% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-ed50-analysis.R
\name{fit_drms}
\alias{fit_drms}
\title{Fit Dynamic Regression Models (DRMs)}
\usage{
fit_drms(dataset, grouping_properties, drm_formula)
}
\arguments{
\item{dataset}{A data frame containing the dataset on which to fit the DRMs.}

\item{grouping_properties}{A character vector specifying the names of columns in the dataset that will be used as grouping properties for fitting separate DRMs.}

\item{drm_formula}{A formula specifying the dynamic regression model to be fitted. This formula should follow the standard R formula syntax (e.g., y ~ x1 + x2).}
}
\value{
A list of fitted DRM models, with each element corresponding to a unique combination of grouping property values.
}
\description{
This function fits dynamic regression models (DRMs) to a given dataset using the specified grouping properties and DRM formula.
}
\examples{
fit_drms(data, c("Site", "Condition", "Species", "Genotype"), "PAM ~ Temperature")

}
\keyword{modeling}
