% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-and-validate-data.R
\name{preprocess_dataset}
\alias{preprocess_dataset}
\title{Preprocesses the data by converting and checking column data types.}
\usage{
preprocess_dataset(dataset)
}
\arguments{
\item{dataset}{A data frame containing the dataset to be preprocessed.}
}
\value{
A preprocessed data frame with converted and validated column data types.
}
\description{
This function preprocesses the input data by performing checks on column data types
and converting them if necessary. It ensures that the dataset meets certain requirements
before further analysis or modeling.
}
\examples{
# Load a sample dataset
data(cbass_dataset)
# Preprocess the dataset
preprocessed_data <- preprocess_dataset(cbass_dataset)
}
