% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{process_and_validate_cbass_dataset}
\alias{process_and_validate_cbass_dataset}
\title{Process and Validate CBASS Dataset}
\usage{
process_and_validate_cbass_dataset(dataset)
}
\arguments{
\item{dataset}{A data frame representing the CBASS dataset to be processed and validated.}
}
\value{
A processed and validated CBASS dataset with appropriate data types for its columns.
}
\description{
This function processes and validates a dataset to ensure it contains all the mandatory columns required for further processing.
If any mandatory columns are missing, it raises an error with the list of missing columns.
Additionally, the function checks and converts the columns in the dataset.
}
\examples{
data(cbass_dataset) # Assuming a dataset named 'cbass_dataset' is available in the environment
processed_dataset <- process_and_validate_cbass_dataset(cbass_dataset)

}
\seealso{
\code{\link{dataset_has_mandatory_columns}}, \code{\link{check_and_convert_columns}}
}
