% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-and-validate-data.R
\name{read_data}
\alias{read_data}
\title{Read Data Function}
\usage{
read_data(file_path)
}
\arguments{
\item{file_path}{Character string specifying the path to the input file.}
}
\value{
A data frame containing the read data.
}
\description{
Reads data from a file based on its format (Excel or CSV).
}
\details{
This function determines the file format based on the file extension and uses appropriate methods to read data from either Excel (xls, xlsx) or CSV (csv, txt) files.
}
\examples{
# Read data from an Excel file
read_data(system.file("extdata", "example.xlsx", package = "CBASSED50"))

# Read data from a CSV file
read_data(system.file("extdata", "example.csv", package = "CBASSED50"))

}
