% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature-responce-curve.R
\name{transform_predictions_to_long_dataframe}
\alias{transform_predictions_to_long_dataframe}
\title{Transform Predictions to a Long-Format DataFrame}
\usage{
transform_predictions_to_long_dataframe(predictions, temp_range)
}
\arguments{
\item{predictions}{A list of data frames where each data frame represents
predictions for a specific temperature range. The data frames should have
a common grouping property.}

\item{temp_range}{the temperature range to be used for predictions from the function define_temperature_ranges}
}
\value{
A long-format data frame containing the transformed predictions with
  columns for "GroupingProperty," "Temperature," and "PredictedPAM."
}
\description{
This function takes a list of predictions and converts them into a long-format
data frame. Each prediction corresponds to a different temperature range.
}
