% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbass_dataset}
\alias{cbass_dataset}
\title{CBASS Dataset}
\format{
A data frame with 300 observations and 9 variables:
  \describe{
    \item{Date}{A character vector representing the date of the observation.}
    \item{Country}{A character vector representing the country of the observation.}
    \item{Site}{A character vector representing the site of the observation.}
    \item{Condition}{A character vector representing the treatment or condition of the observation.}
    \item{Species}{A character vector representing the species of the observation.}
    \item{Genotype}{An integer vector representing the genotype of the observation.}
    \item{Temp}{An integer vector representing the temperature of the observation.}
    \item{Timepoint}{An integer vector representing the timepoint of the observation.}
    \item{PAM}{A numeric vector representing the PAM value of the observation.}
  }
}
\usage{
data(cbass_dataset)
}
\description{
A dataset containing example experimental data.
}
\details{
This dataset provides experimental data with various attributes for
demonstration and testing purposes.
}
\examples{
# Load the sample dataset
data(cbass_dataset)
head(cbass_dataset)

}
\keyword{datasets}
